택시 동승 가능여부 변경, 신청 및 취소 API 연동

+++ app/src/main/java/com/example/saengsaengtalk/APIS/DataModels/Taxi.kt
@@ -13,7 +13,7 @@ data class TaxiPostingModel(
 
 /** 303 택시 게시글 조회 모델 */
 data class TaxiPostModel(
-    val user_id: Int,
+    val user_id: Long,
     val nick_name: String,
     val title: String,
     val content: String?,
